Add files via upload
This project is inspired by the Joint-based Muscle Deformation technology used in God of War: Ragnarok, introduced by Tenghao Wang at GDC talk. Comparing to traditional muscle simulation technology, this pipeline offers perfect in-engine realistic deformation effect. Based on the pipeline, I developed a tool based on Maya python, which allows user to create a group in several seconds. 
Each muscle is driven by a joint group, which contains several joints. The muscle origin joint and muscle insertion joint are attached to corresponding parts of the bones. Aim constraints are created to simulate muscle attachments, and point constraints are for muscle movements.
To preserve the muscles' volume while the bone moves and create the bulge/stretch effect, the system will calculate the length change of the muscle, and modify the scale value of muscle bind joint via setting drivien keys
The python tool allows user to assign the attach joints of muscle group, the stretch factor and compress factor. In Edit mode, the user can edit the locations of muscle origin and insertion to customize the micro offsets based on anatomy knowledge.